+++ src/clients/sens/sms.ts
@@ -1,18 +1,47 @@
 import { Method } from "axios"
 import { NCPAuthKeyType, SMSserviceAuthType } from "../../types/auth_types"
 import { SendSMSParamType } from "../../types/param_types"
-import { ApiClientResponse, SearchMessageRequestReturnType, SearchMessageResultReturnType, SendSMSReturnType } from "../../types/return_types"
+import { ApiClientResponse, SendSMSReturnType, SearchMessageRequestReturnType, SearchMessageResultReturnType } from "../../types/return_types"
+import { ApiClient, ApiRequest } from "../../utils/api_request"
 import { generateApiSignature } from "../../utils/helper"
-import { ApiRequest, MockApiClient } from "./mock_apiClient"
 
-
-export class MockSMS {
-
-  private client: MockApiClient
+export class SMS {
+  /**
+   * The account access Key for API authentication
+   * 
+   * @access private 
+   * @type {ApiClient} 
+   * @memberof SMS
+   */
+  private client: ApiClient
   
+  
+  /**
+   * The account access Key for API authentication
+   * 
+   * @access private 
+   * @type {NCPAuthKeyType} 
+   * @memberof SMS
+   */
   private ncpAuthKey: NCPAuthKeyType
+
+  /**
+   * The phoneNumber, serviceId for using SMS API
+   * 
+   * @access private 
+   * @type {SMSserviceAuthType} 
+   * @memberof SMS
+   */
   private smsAuth: SMSserviceAuthType
 
+  /**
+   * Creates an instance of SMS service agent.
+   * 
+   * @param {string} baseURL The baseURL of each session ( based on service )
+   * @param {NCPAuthKeyType} ncpAuthKey NCP API authentication data for using API
+   * @param {SMSserviceAuthType} smsAuth Service identification for using SMS Service
+   * @memberof SMS
+   */
   constructor(
     baseUrl: string,
     ncpAuthKey: NCPAuthKeyType,
@@ -20,9 +49,19 @@ export class MockSMS {
   ) {
     this.smsAuth = smsAuth
     this.ncpAuthKey = ncpAuthKey
-    this.client = new MockApiClient(ncpAuthKey, baseUrl)
+    this.client = new ApiClient(ncpAuthKey, baseUrl)
+    
   }
-
+  /**
+   * Construct SendSMS Service apiRequest. Generate signature for signing, pass SendSMSParams to body.
+   * Single / Multiple send mode supported.
+   * @async
+   * @access public
+   * @param {(SendSMSParamType | SendSMSParamType[])} apiRequest - Message-reference Data with ( to, content ). Single / Multi value passing supported.
+   * @param {(number | undefined)} countryCode - CountryCode for recipient. ( default : 82 )
+   * @returns {Promise<ApiClientResponse<SendSMSReturnType>>} return Promise response of http request with current ApiRequest configs and handle errors
+   * @memberof SMS
+   */
   public async sendSMS(smsParam: SendSMSParamType | SendSMSParamType[], countryCode?: number): Promise<ApiClientResponse<SendSMSReturnType>> {
     // construct path and method
     const path = `/sms/v2/services/${this.smsAuth.serviceId}/messages`
@@ -36,11 +75,12 @@ export class MockSMS {
       'x-ncp-apigw-signature-v2': signature,
     }
     // construct message parameters and body
+    countryCode = (countryCode === undefined) ? 82 : countryCode
     smsParam = (smsParam instanceof Array) ? smsParam : [smsParam]
     const body = {
       'type': 'SMS',
       'contentType': 'COMM',
-      'countryCode': countryCode,
+      'countryCode': `${countryCode}`,
       'from': this.smsAuth.phone,
       'content': ' ',
       'messages': smsParam
@@ -54,6 +94,15 @@ export class MockSMS {
     return this.client.request<SendSMSReturnType>(apiRequest)
   }
 
+  /**
+   * Construct searchMessageRequest Service apiRequest. Generate signature for signing, pass requestId for search.
+   * return the requestData for given requestId. It also returns messages info refer to the request.
+   * @async
+   * @access public
+   * @param {string} requestId - requestId which get from response of sendSMS API
+   * @returns {Promise<ApiClientResponse<SearchMessageRequestReturnType>>} return Promise response of http request with current ApiRequest configs and handle errors
+   * @memberof SMS
+   */
   public async searchMessageRequest(requestId: string): Promise<ApiClientResponse<SearchMessageRequestReturnType>> {
     // construct path and method
     const path = `/sms/v2/services/${this.smsAuth.serviceId}/messages?requestId=${requestId}`
@@ -74,6 +123,16 @@ export class MockSMS {
     return this.client.request<SearchMessageRequestReturnType>(apiRequest)
   }
 
+  /**
+   * Construct searchMessageResult Service apiRequest. Generate signature for signing, pass requestId for search.
+   * return the requestData for given messageId. It also returns messages info refer to the request.
+   * It's response data is more detail than searchMessageRequest with completeTime, statuses, telcoCode, etc.
+   * @async
+   * @access public
+   * @param {string} messageId - messageId which get from response of sendSMS API
+   * @returns {Promise<ApiClientResponse<SearchMessageRequestReturnType>>} return Promise response of http request with current ApiRequest configs and handle errors
+   * @memberof SMS
+   */
   public async searchMessageResult(messageId: string): Promise<ApiClientResponse<SearchMessageResultReturnType>> {
     // construct path and method
     const path = `/sms/v2/services/${this.smsAuth.serviceId}/messages/${messageId}`
@@ -94,5 +153,5 @@ export class MockSMS {
     return this.client.request<SearchMessageResultReturnType>(apiRequest)
   }
 
-
+  
 }